# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API для Todo и Auth",
    "version": "1.0.0",
    "description": "Документация API для управления задачами и пользователями"
  },
  "servers": [
    {
      "url": "https://expresslesson1-last.onrender.com/api/v1",
      "description": "Продакшн сервер"
    },
    {
      "url": "http://localhost:5000/api/v1",
      "description": "local сервер"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/todo/getByCategory/{category}": {
      "get": {
        "summary": "Получить задачи по категории",
        "operationId": "getProductsByCategory",
        "tags": ["Todo"],
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "GADGETS",
                "TECHNICS",
                "CLOTHES_MEN",
                "CLOTHES_WOMEN",
                "CONSTRUCTION_EQUIPMENT",
                "BOOKS"
              ]
            },
            "description": "Категория товара, по которой нужно получить задачи"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Список задач по категории",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "title": {
                            "type": "string"
                          },
                          "price": {
                            "type": "number"
                          },
                          "description": {
                            "type": "string"
                          },
                          "image": {
                            "type": "string"
                          },
                          "authorEmail": {
                            "type": "string"
                          },
                          "isCompleted": {
                            "type": "boolean"
                          },
                          "category": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "author": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer"
                              },
                              "name": {
                                "type": "string"
                              },
                              "email": {
                                "type": "string"
                              },
                              "profilePhoto": {
                                "type": "string"
                              },
                              "isAdmin": {
                                "type": "boolean"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}